@@ -14,33 +14,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.4",
-    "@angular/cdk": "^7.1.4",
-    "@angular/common": "^7.1.4",
-    "@angular/compiler": "^7.1.4",
-    "@angular/core": "^7.1.4",
+    "@angular/animations": "^7.2.5",
+    "@angular/cdk": "^7.3.2",
+    "@angular/common": "^7.2.5",
+    "@angular/compiler": "^7.2.5",
+    "@angular/core": "^7.2.5",
     "@angular/fire": "^5.1.1",
-    "@angular/forms": "^7.1.4",
-    "@angular/http": "^7.1.4",
-    "@angular/material": "^7.2.0",
-    "@angular/platform-browser": "^7.1.4",
-    "@angular/platform-browser-dynamic": "^7.1.4",
-    "@angular/router": "^7.1.4",
+    "@angular/forms": "^7.2.5",
+    "@angular/http": "^7.2.5",
+    "@angular/material": "^7.3.2",
+    "@angular/platform-browser": "^7.2.5",
+    "@angular/platform-browser-dynamic": "^7.2.5",
+    "@angular/router": "^7.2.5",
+    "@ngrx/effects": "^7.2.0",
+    "@ngrx/entity": "^7.2.0",
+    "@ngrx/store": "^7.2.0",
+    "@ngrx/store-devtools": "^7.2.0",
     "aws-sdk": "^2.377.0",
     "core-js": "^2.6.1",
     "firebase": "^5.7.2",
     "firebase-admin": "^6.4.0",
     "firebase-functions": "^2.1.0",
     "firebase-tools": "^6.2.2",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.4.0",
     "tslib": "^1.9.3",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.11.4",
-    "@angular/cli": "~7.1.4",
-    "@angular/compiler-cli": "^7.1.4",
-    "@angular/language-service": "^7.1.4",
+    "@angular-devkit/build-angular": "~0.13.1",
+    "@angular/cli": "~7.3.1",
+    "@angular/compiler-cli": "^7.2.5",
+    "@angular/language-service": "^7.2.5",
+    "@ngrx/schematics": "^7.2.0",
     "@types/jasmine": "~3.3.5",
     "@types/jasminewd2": "~2.0.5",
     "@types/node": "~10.12.0",
@@ -53,10 +58,10 @@
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.3.1",
     "protractor": "~5.4.2",
-    "stylelint": "^9.9.0",
+    "stylelint": "^9.10.1",
     "stylelint-config-standard": "^18.2.0",
     "ts-node": "~7.0.1",
-    "tslint": "~5.12.0",
-    "typescript": "~3.1"
+    "tslint": "~5.12.1",
+    "typescript": "3.2"
   }
 }
