@@ -4,10 +4,12 @@ import { AuthService } from '../../auth/auth.service';
 import { Observable, of } from 'rxjs';
 import { MatDialog } from '@angular/material';
 import { CreateGroupDialogComponent } from '../create-group-dialog/create-group-dialog.component';
-import { map } from 'rxjs/operators';
-import { User } from 'firebase';
+import { map, concatMap, concat, mergeMap } from 'rxjs/operators';
 import { GroupService } from 'src/app/services/group/group.service';
 import { Router } from '@angular/router';
+import { Store, select } from '@ngrx/store';
+import { selectUser } from 'src/app/store/reducers/user/user.reducer';
+import { User } from 'src/app/models/user.model';
 
 @Component({
   selector: 'app-group-list',
@@ -15,15 +17,34 @@ import { Router } from '@angular/router';
   styleUrls: ['./group-list.component.scss']
 })
 export class GroupListComponent implements OnInit {
-  user: User | null;
-  items: Observable<any> = of([]);
-  userDoc: AngularFirestoreDocument;
+  user?: User;
+  user$: Observable<User>;
+  items$: Observable<any> = of([]);
+  userDoc?: AngularFirestoreDocument;
 
-  constructor(private db: AngularFirestore, private authService: AuthService,
+  constructor(private db: AngularFirestore,
+    private store: Store<any>,
     private dialog: MatDialog, private router: Router, private groupService: GroupService) {
-    this.user = this.authService.getUser();
-    this.userDoc = this.db.doc(`/users/${this.user ? this.user.uid : ''}`);
-    this.items = this.userDoc.collection('/groups').snapshotChanges().pipe(
+    this.user$ = this.store.pipe(select(selectUser));
+
+    this.user$.pipe(
+      map(user => {
+        if (!user) {
+          throw new Error('not authenticated');
+        }
+        this.user = user;
+        this.userDoc = this.db.doc(`/users/${user.uid}`);
+        this.items$ = this.getItems();
+      })
+    ).subscribe();
+  }
+
+  ngOnInit() {
+  }
+
+  private getItems() {
+    if (!this.userDoc) { return of([]); }
+    return this.userDoc.collection('/groups').snapshotChanges().pipe(
       map(actions => {
         return actions.map(item => {
           return {
@@ -33,9 +54,6 @@ export class GroupListComponent implements OnInit {
         });
       })
     );
-  }
-
-  ngOnInit() {
   }
 
   openCreateDialog() {
