@@ -11,7 +11,7 @@ const routes: Routes = [
 
 @NgModule({
   imports: [RouterModule.forRoot(routes, {
-    enableTracing: true
+    enableTracing: false
   })],
   exports: [RouterModule],
   providers: []
