@@ -4,6 +4,11 @@ import { auth } from 'firebase';
 import { FormControl, Validators, FormGroup } from '@angular/forms';
 import { Router } from '@angular/router';
 import { SocialButton } from 'src/app/enums/social-button.enum';
+import { Store, select } from '@ngrx/store';
+import * as authActions from 'src/app/actions/auth.actions';
+import { selectUser, UserState, selectIsUserLoggedIn } from 'src/app/store/reducers/user/user.reducer';
+import { UserActionTypes } from 'src/app/store/actions/user/user.actions';
+import * as userActions from 'src/app/store/actions/user/user.actions';
 
 @Component({
   selector: 'app-login',
@@ -15,7 +20,8 @@ export class LoginComponent implements OnInit {
   password: FormControl;
   isSignIn = true;
 
-  constructor(private afAuth: AngularFireAuth, private router: Router) {
+  constructor(private afAuth: AngularFireAuth, private router: Router,
+    private store: Store<UserState>) {
     this.email = new FormControl('', [
       Validators.required,
       Validators.email
@@ -28,6 +34,13 @@ export class LoginComponent implements OnInit {
   }
 
   ngOnInit() {
+    this.store.pipe(select(selectIsUserLoggedIn)).subscribe(loggedIn => {
+      console.log('user loggedIn: ', loggedIn);
+      if (loggedIn) {
+        this.router.navigate(['/home/groups'], { replaceUrl: true });
+      }
+    });
+    this.store.dispatch(new userActions.GetUser());
   }
 
   login() {
@@ -37,8 +50,8 @@ export class LoginComponent implements OnInit {
   }
 
   async googleLogin() {
-    await this.afAuth.auth.signInWithPopup(new auth.GoogleAuthProvider());
-    this.router.navigate(['/home/groups']);
+    this.store.dispatch(new authActions.GoogleLogin());
+    // this.router.navigate(['/home/groups']);
   }
 
   signInWithSocial(account: SocialButton) {
